Extract Pool class to the separate file
