+++ metautil.js
@@ -3,5 +3,6 @@
 const utilities = require('./lib/utilities.js');
 const secutity = require('./lib/security.js');
 const semaphore = require('./lib/semaphore.js');
+const pool = require('./lib/pool.js');
 
-module.exports = { ...utilities, ...secutity, ...semaphore };
+module.exports = { ...utilities, ...secutity, ...semaphore, ...pool };
